// src/components/screens/ownership-form-step/ownership-form-step.interface.ts
export interface OwnershipStepForm {
	activity: string;
	itn?: string;
	itnScan?: string;
	registerDate?: string;
	psrnst?: string;
	psrnstScan?: string;
	leaseСontractScan?: string;
	stateRegisterSTScan?: string;
	noContract?: boolean;
	fullCompanyName?: string;
	shortCompanynName?: string;
	psrn?: string;
	psrnScan?: string;
}